requirements.txt
torch
datasets
transformers
numpy
wandb
python-dotenv
nvidia-ml-py